+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.0 - 2026-02-18
+
+### Changed
+
+- Upgrade `:restlax` to `1.0.0`
+- Consolidate auth usage around `Cloudflare.Client.new/1` and `Cloudflare.Client.init/1`
+
 ## 0.2.0 - 2021-08-24
 
 ### Fixed